- Correction du temps en double - Ajout du SimulateurTaxis.java

+++ src/simulateur/GenerateurTempsAttente.java
@@ -1,5 +1,5 @@
 package simulateur;
 
 public interface GenerateurTempsAttente {
-	public int genererTempsAttente();
+	public double genererTempsAttente();
 }
